+++ extensions/luna-search/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Luna Search Changelog
 
-## [Bug Fix & Optimizations] - 2024-08-24
+## [Menu Bar Command, Bug Fixes & Optimizations] - 2024-08-24
 
+- Added a new menu-bar command that supports quickly accessing the most recent games the user interacted with and the top 5 trending games.
 - Noticed a bug with the late-breaking pagination feature - this is fixed! No more errors.
 - Internal refactoring should make Luna Search a little more snappy and a little less janky.